+++ llama_index/llms/__init__.py
@@ -1,15 +1,39 @@
-from llama_index.llms.base import ChatMessage, ChatResponse, ChatResponseGen
+from llama_index.llms.anthropic import Anthropic
+from llama_index.llms.azure_openai import AzureOpenAI
+from llama_index.llms.base import (
+    ChatMessage,
+    ChatResponse,
+    ChatResponseAsyncGen,
+    ChatResponseGen,
+    CompletionResponse,
+    CompletionResponseAsyncGen,
+    CompletionResponseGen,
+    LLMMetadata,
+    MessageRole,
+)
+from llama_index.llms.custom import CustomLLM
 from llama_index.llms.huggingface import HuggingFaceLLM
 from llama_index.llms.langchain import LangChainLLM
+from llama_index.llms.mock import MockLLM
 from llama_index.llms.openai import OpenAI
 from llama_index.llms.palm import PaLM
 
 __all__ = [
     "OpenAI",
+    "AzureOpenAI",
     "LangChainLLM",
     "HuggingFaceLLM",
+    "PaLM",
+    "Anthropic",
+    "CustomLLM",
+    "MockLLM",
     "ChatMessage",
+    "MessageRole",
     "ChatResponse",
     "ChatResponseGen",
-    "PaLM",
+    "ChatResponseAsyncGen",
+    "CompletionResponse",
+    "CompletionResponseGen",
+    "CompletionResponseAsyncGen",
+    "LLMMetadata",
 ]